requirements.txt
playwright>=1.40.0
playwright-stealth>=2.0.0
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
